pdc: tps6699x: Set pmc from devicetree
..so we can use unified firmware.

BUG=b:388916590
TEST=(1) ./twister -T zephyr/test/pdc
TEST=(2) Deploy firmware to a 2xTPS66993 device
TEST=(3) In ec console, see that one of TPS66993 pmc address is updated
         with log "Set PMC address ${APPCONFIG_PORT} -> ${DT_PORT}"
TEST=(4) Insert a dongle and see
         "PD status updated recovery details: 0x23"
TEST=(5) Run "pdc reset 0" or "pdc reset 1"
TEST=(6) "Set PMC address ${APPCONFIG_PORT} -> ${DT_PORT}" shows again

+++ zephyr/include/drivers/pdc.h
@@ -107,7 +107,7 @@ struct pdc_info_t {
 	/** Set to 1 if running from flash code (optional) */
 	uint8_t is_running_flash_code;
 	/** Set to the currently used flash bank (optional) */
-	uint8_t running_in_flash_bank;
+	uint32_t running_in_flash_bank;
 	/** 12-byte program name string plus NUL terminator */
 	char project_name[USB_PD_CHIP_INFO_PROJECT_NAME_LEN + 1];
 	/** Compat string of driver */
